Added so that you always start where you end and debug

+++ app/src/main/java/com/example/yournal/SettingsFragment.kt
@@ -1,5 +1,6 @@
 package com.example.yournal
 
+import android.app.AlertDialog
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -23,7 +24,18 @@ class SettingsFragment : Fragment() {
         val view = binding.root
 
         binding.btnDeleteAll.setOnClickListener {
-            TripManager.deleteAll(requireContext())
+            AlertDialog.Builder(requireContext())
+                .setTitle("Delete all")
+                .setMessage("Are you sure you want to delete all?")
+                .setPositiveButton("Yes") { dialog, _ ->
+                    TripManager.deleteAll(requireContext())
+                    dialog.dismiss() // Close the dialog
+                }
+                .setNegativeButton("No") { dialog, _ ->
+                    dialog.dismiss() // Close the dialog
+                }
+                .show()
+
         }
         // Inflate the layout for this fragment
         return view
